refactor: extract function to file
